requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-Migrate
Flask-WTF
Authlib
requests
python-dotenv
